https://www.figma.com/file/eX7dxzmkAMKjiZYToLs4sW/Paytm-Copy-paste?type=design&node-id=0%3A1&mode=design&t=LX7AuclbkcPgvydr-1
This is a Figma File Designed for Paytm
It was designed on Figma, which shows the working of the Paytm payment application.
All the screens have a working prototype.
Paytm is an online payment application that helps users to make ONLINE PaYMENTS
So I have designed it for a better understanding of the Design tool Figma. Where I used the concepts of Autolayout prototype and wireframing.